refactor theam
refactored app name of spring project
