+++ starterkit/src/main/java/master/nigaits/wapp/config/ThemeConfig.java
@@ -1,9 +1,9 @@
-package com.theme.starterkit.config;
+package master.nigaits.wapp.config;
 
-import com.theme.starterkit.libs.KTTheme;
-import com.theme.starterkit.libs.config.KTThemeBaseConfig;
-import com.theme.starterkit.libs.config.KTIconsBaseSettings;
-import com.theme.starterkit.libs.config.KTThemeSettings;
+import master.nigaits.wapp.libs.KTTheme;
+import master.nigaits.wapp.libs.config.KTThemeBaseConfig;
+import master.nigaits.wapp.libs.config.KTIconsBaseSettings;
+import master.nigaits.wapp.libs.config.KTThemeSettings;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.context.annotation.RequestScope;
