+++ starterkit/src/main/java/master/nigaits/wapp/libs/KTBootstrap.java
@@ -1,6 +1,6 @@
-package com.theme.starterkit.libs;
+package master.nigaits.wapp.libs;
 
-import com.theme.starterkit.libs.config.KTThemeBaseConfig;
+import master.nigaits.wapp.libs.config.KTThemeBaseConfig;
 
 public class KTBootstrap {
