+++ starterkit/src/main/java/master/nigaits/wapp/libs/config/KTThemeAssets.java
@@ -1,4 +1,4 @@
-package com.theme.starterkit.libs.config;
+package master.nigaits.wapp.libs.config;
 
 import java.util.List;
